ICSharpCode.SharpZipLib implemented, plugin detection
ModularLauncherUtil now has a wrapper for ICSharpCode.SharpZipLib to
unzip files, zipping to be done at a later stage.
Launcher will now scan for plugins and give out a warning to the user if
none can be found. Warning UserControl could later on contain useful
information on how to make or where to get plugins.