Settings set up
Started work on the Settings; it'll now take the config from the main
thread and distribute the bits to the individual plugins it loads.
Plugins will have to implement a function to supply the information to
their settings and update them accordingly. I should probably move that
function to the settings panel itself, though.